requirements.txt
Babel==2.4.0
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.4
django-allauth==0.33.0
django-modeltranslation==0.12.1
django-mptt==0.8.7
django-multiselectfield==0.1.8
django-phonenumber-field==1.3.0
django-widget-tweaks==1.4.1
django-wysiwyg-redactor==0.5.1
idna==2.6
mysqlclient==1.3.10
oauthlib==2.0.2
olefile==0.44
phonenumberslite==8.7.1
Pillow==4.2.1
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
sorl-thumbnail==12.3
urllib3==1.22
